# OrderLines
- Only Resources for the OrderLines table were created
- It is worth mentioning that there is no direct access to OrderLine.  When the Order is accessed it allows specifying the "includeLines=true" parameter to get the related order lines.

+++ app/Http/Controllers/Api/V1/OrderController.php
@@ -10,6 +10,7 @@ use App\Http\Resources\V1\OrderCollection;
 use App\Http\Resources\V1\OrderResource;
 use App\Models\Customer;
 use App\Models\Order;
+use App\Models\OrderLine;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -21,13 +22,13 @@ class OrderController extends Controller
         $filter = new OrderFilter();
         $filterItems = $filter->transform($request);
 
-        // $includeCustomers = $request->query('includeCustomers');
+        $includeLines = $request->query('includeLines');
 
         $orders = Order::where($filterItems);
 
-        // if ($includeCustomers) {
-        //     $orders->with('customers');
-        // }
+        if ($includeLines) {
+            $orders->with('orderLines');
+        }
         return new OrderCollection($orders->paginate()->appends($request->query()));
 
         // $orders = Order::paginate();
@@ -36,18 +37,17 @@ class OrderController extends Controller
 
     public function store(OrderStoreRequest $request): OrderResource
     {
-        info('Store');
         $order = Order::create($request->validated());
-        info('Created');
+
         return new OrderResource($order);
     }
 
     public function show(Request $request, Order $order): OrderResource
     {
-        $includeCustomers = $request->query('includeCustomers');
+        $includeLines = $request->query('includeLines');
 
-        if ($includeCustomers) {
-            $order->loadMissing('customers');
+        if ($includeLines) {
+            $order->loadMissing('orderLines');
         }
 
         return new OrderResource($order);
@@ -62,12 +62,12 @@ class OrderController extends Controller
 
     public function destroy(Request $request, Order $order)
     {
-        // $intCustQnty = Customer::where('Order_id', $order->id)->get()->count();
-        // if ($intCustQnty == 0) {
+        $intLineQnty = OrderLine::where('order_id', $order->id)->get()->count();
+        if ($intLineQnty == 0) {
             $order->delete();
             return response()->noContent();
-        // } else {
-        //     return response()->json(['errors' => 'There are ('. $intCustQnty. ') customers assigned to this Order.'], 400);
-        // }
+        } else {
+            return response()->json(['errors' => 'There are ('. $intLineQnty. ') order lines assigned to this Order.'], 400);
+        }
     }
 }
